+++ src/awtk_global.c
@@ -181,15 +181,10 @@ ret_t tk_init(wh_t w, wh_t h, app_type_t app_type, const char* app_name, const c
   loop = main_loop_init(w, h);
   return_value_if_fail(loop != NULL, RET_FAIL);
 
-  WINDOW_MANAGER(window_manager())->canvas = &((loop)->canvas);
-
   return RET_OK;
 }
 
 ret_t tk_deinit_internal(void) {
-  widget_destroy(window_manager());
-  window_manager_set(NULL);
-
   clip_board_destroy(clip_board());
   clip_board_set(NULL);
 
@@ -204,10 +199,16 @@ ret_t tk_deinit_internal(void) {
 
   children_layouter_factory_destroy(children_layouter_factory());
   children_layouter_factory_set(NULL);
+  
+  image_manager_destroy(image_manager());
+  image_manager_set(NULL);
+
+  widget_destroy(window_manager());
+  window_manager_set(NULL);
 
   widget_animator_manager_destroy(widget_animator_manager());
   widget_animator_manager_set(NULL);
-
+  
   idle_manager_destroy(idle_manager());
   idle_manager_set(NULL);
 
@@ -216,15 +217,12 @@ ret_t tk_deinit_internal(void) {
 
   widget_factory_destroy(widget_factory());
   widget_factory_set(NULL);
-
+  
   input_method_destroy(input_method());
   input_method_set(NULL);
 
   theme_destroy(theme());
   theme_set(NULL);
-
-  image_manager_destroy(image_manager());
-  image_manager_set(NULL);
 
   font_manager_destroy(font_manager());
   font_manager_set(NULL);
@@ -276,41 +274,24 @@ ret_t tk_set_lcd_orientation(lcd_orientation_t orientation) {
                        RET_NOT_IMPL);
 
   if (info->lcd_orientation != orientation) {
-    wh_t w = info->lcd_w;
-    wh_t h = info->lcd_h;
-    lcd_t* lcd = loop->canvas.lcd;
+    orientation_event_t e;
+    orientation_event_init(&e, EVT_ORIENTATION_WILL_CHANGED, NULL, orientation);
+    widget_dispatch(window_manager(), (event_t*)&e);
 
     system_info_set_lcd_orientation(info, orientation);
-    if (orientation == LCD_ORIENTATION_90 || orientation == LCD_ORIENTATION_270) {
-      w = info->lcd_h;
-      h = info->lcd_w;
-    }
-
-    lcd_resize(lcd, w, h, 0);
-
-    window_manager_resize(window_manager(), w, h);
   }
 
   return RET_OK;
 }
 
 int32_t tk_get_pointer_x(void) {
-  window_manager_t* wm = WINDOW_MANAGER(window_manager());
-  return_value_if_fail(wm != NULL, 0);
-
-  return wm->input_device_status.last_x;
+  return window_manager_get_pointer_x(window_manager());
 }
 
 int32_t tk_get_pointer_y(void) {
-  window_manager_t* wm = WINDOW_MANAGER(window_manager());
-  return_value_if_fail(wm != NULL, 0);
-
-  return wm->input_device_status.last_y;
+  return window_manager_get_pointer_y(window_manager());
 }
 
 bool_t tk_is_pointer_pressed(void) {
-  window_manager_t* wm = WINDOW_MANAGER(window_manager());
-  return_value_if_fail(wm != NULL, 0);
-
-  return wm->input_device_status.pressed;
+  return window_manager_get_pointer_pressed(window_manager());
 }