+++ src/main_loop/main_loop_simple.c
@@ -161,7 +161,7 @@ static ret_t main_loop_simple_step(main_loop_t* l) {
   main_loop_dispatch_events(loop);
   idle_dispatch();
 
-  window_manager_paint(loop->base.wm, &(loop->base.canvas));
+  window_manager_paint(loop->base.wm);
 
   return RET_OK;
 }
@@ -199,7 +199,7 @@ main_loop_simple_t* main_loop_simple_init(int w, int h) {
   loop->base.step = main_loop_simple_step;
   loop->base.queue_event = main_loop_simple_queue_event;
 
-  window_manager_resize(loop->base.wm, w, h);
+  window_manager_post_init(loop->base.wm, w, h);
   main_loop_set((main_loop_t*)loop);
 
   return loop;
@@ -209,9 +209,7 @@ ret_t main_loop_simple_reset(main_loop_simple_t* loop) {
   return_value_if_fail(loop != NULL, RET_BAD_PARAMS);
   event_queue_destroy(loop->queue);
   tk_mutex_destroy(loop->mutex);
-  lcd_destroy(loop->base.lcd);
 
-  canvas_reset(&(loop->base.canvas));
   memset(loop, 0x00, sizeof(main_loop_simple_t));
 
   return RET_OK;